Use Microstation bridge to process ifc files. Better logging messages for point cloud upload failures. Clean up of header files

Related work items: #107465

+++ Bridges/Civil/ORDBridge/Tests/BackDoorORDBridge/ORDBridgeTestsFixture.cpp
@@ -1148,7 +1148,7 @@ DgnDbPtr CiviliModelBridgesORDBridgeTestsFixture::VerifyConvertedElementItemType
 
     BentleyApi::ECN::ECClassCP ecClass = dynSchema->GetClassCP(typeClassName);
     BeAssert(ecClass != NULL);
-    int classPropCount = ecClass->GetPropertyCount();
+    size_t classPropCount = ecClass->GetPropertyCount();
     BeAssert(typePropCount == classPropCount);
 
     ECSqlStatement stmt;
